FAC-569:feature:Move GetNotificationPlatformSettings classes to notification package
