+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/notification/GetNotificationPlatformSettingsService.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.users.application.service.user;
+package org.flickit.assessment.users.application.service.notification;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.application.domain.notification.NotificationSenderSettingProvider;
-import org.flickit.assessment.users.application.port.in.user.GetNotificationPlatformSettingsUseCase;
+import org.flickit.assessment.users.application.port.in.notification.GetNotificationPlatformSettingsUseCase;
 import org.springframework.stereotype.Service;
 
 @Service
