+++ flickit-assessment-users/src/test/java/org/flickit/assessment/users/application/port/in/notification/GetNotificationPlatformSettingsUseCaseParamTest.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.users.application.port.in.user;
+package org.flickit.assessment.users.application.port.in.notification;
 
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.Test;
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 class GetNotificationPlatformSettingsUseCaseParamTest {
 
     @Test
-    void testGetUserSubscriberHash_CurrentUserIdIsNull_ErrorMessage() {
+    void testGetNotificationPlatformSettings_CurrentUserIdIsNull_ErrorMessage() {
         var throwable = assertThrows(ConstraintViolationException.class,
             () -> new GetNotificationPlatformSettingsUseCase.Param(null));
         assertThat(throwable).hasMessage("currentUserId: " + COMMON_CURRENT_USER_ID_NOT_NULL);
